+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file. This projec
 - fix: apply per-device config overrides to devices (#83) (@colbyr)
 - style: fix lint issues in the beta changes
 - fix: use url resolution when following authentication redirects (#100) (@smitty078)
+- feat: add air conditioner default mode, separate fan service and dry switch options (#101) (@smitty078)
 
 ## [0.4.16](https://github.com/homebridge-plugins/homebridge-smarthq/compare/v0.4.15...v0.4.16) (2026-04-09)